Update Task priority type from string to RequestPriority
We have a minor data integrity issue where some of our tasks have priorities with incorrect casing, which our types allow since it is typed as a string. This PR updates it to use the enum we use in our core repo to help avoid this data skew issue.  Note that we are breaking the spec with this somewhat, though it doesn't really matter since we don't have a real need for being compliant with the fhir spec for tasks, they're internal only.